refactor jsapi - command listener
