+++ src/xprofiler.cc
@@ -1,9 +1,9 @@
-#include "commands/listener.h"
 #include "jsapi/include/export_configure.h"
 #include "jsapi/include/export_environment.h"
 #include "jsapi/include/export_hooks.h"
 #include "jsapi/include/export_http.h"
 #include "jsapi/include/export_logger.h"
+#include "jsapi/include/export_thread_listener.h"
 #include "jsapi/include/export_thread_logbypass.h"
 #include "library/common.h"
 #include "nan.h"
